Update files with pragma once header guards, move main to new file, bump to cxx17.
