+++ CMakeLists.txt
@@ -6,9 +6,9 @@ if(NOT CMAKE_C_STANDARD)
   set(CMAKE_C_STANDARD 99)
 endif()
 
-# Default to C++14
+# Default to C++17
 if(NOT CMAKE_CXX_STANDARD)
-  set(CMAKE_CXX_STANDARD 14)
+  set(CMAKE_CXX_STANDARD 17)
 endif()
 
 if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
@@ -24,7 +24,7 @@ find_package(std_msgs REQUIRED)
 
 # BUILDING
 
-add_executable(lidar_processor src/lidar_processor.cpp)
+add_executable(lidar_processor src/lidar_processor_main.cpp)
 ament_target_dependencies(
   lidar_processor
   pcl_conversions
